app/requirements.txt
aerich==0.5.8; python_version >= "3.7" and python_version < "4.0"
aiosqlite==0.17.0; python_version >= "3.7" and python_version < "4.0"
anyio==3.3.4; python_version >= "3.6" and python_full_version >= "3.6.2"
asgiref==3.4.1; python_version >= "3.6"
asyncpg==0.24.0; python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.6.0"
bcrypt==3.2.0; python_version >= "3.6"
cffi==1.15.0; python_version >= "3.6"
click==8.0.3; python_version >= "3.7" and python_version < "4.0"
colorama==0.4.4; python_version >= "3.6" and python_full_version < "3.0.0" and platform_system == "Windows" or platform_system == "Windows" and python_version >= "3.6" and python_full_version >= "3.5.0"
ddlparse==1.10.0; python_version >= "3.7" and python_version < "4.0"
dictdiffer==0.9.0; python_version >= "3.7" and python_version < "4.0"
ecdsa==0.17.0; python_version >= "2.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0"
fastapi==0.70.0; python_full_version >= "3.6.1"
h11==0.12.0; python_version >= "3.6"
idna==3.3; python_version >= "3.6" and python_full_version >= "3.6.2"
iso8601==0.1.16; python_version >= "3.7" and python_version < "4.0"
passlib==1.7.4
pyasn1==0.4.8; python_version >= "3.5" and python_version < "4"
pycparser==2.20; python_version >= "3.6" and python_full_version < "3.0.0" or python_version >= "3.6" and python_full_version >= "3.4.0"
pydantic==1.8.2; python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.6.1"
pyparsing==3.0.1; python_version >= "3.7" and python_version < "4.0"
pypika-tortoise==0.1.1; python_version >= "3.7" and python_version < "4.0"
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3; python_version >= "3.7" and python_version < "4.0"
rsa==4.7.2; python_version >= "3.5" and python_version < "4"
six==1.16.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_version >= "3.6" and python_full_version >= "3.3.0"
sniffio==1.2.0; python_version >= "3.6" and python_full_version >= "3.6.2"
starlette==0.16.0; python_version >= "3.6" and python_full_version >= "3.6.1"
tortoise-orm==0.17.8; python_version >= "3.7" and python_version < "4.0"
typing-extensions==3.10.0.2; python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.6.1"
uvicorn==0.15.0
